different input process for ed_desc_algo '40
The new process runs significantly faster than before. For some unknown reason, it gives slightly different results than the previous code. They are not better or worse, just different. This should be investigated more thoroughly, but for now I am uploading this version for consistency. The old code is still there, in a multiline string.